// models/ImapServerSearchOptions.ts
/**
 * MailSlurp API
 * MailSlurp is an API for sending and receiving emails and SMS from dynamically allocated email addresses and phone numbers. It's designed for developers and QA teams to test applications, process inbound emails, send templated notifications, attachments, and more.  ## Resources  - [Homepage](https://www.mailslurp.com) - Get an [API KEY](https://app.mailslurp.com/sign-up/) - Generated [SDK Clients](https://docs.mailslurp.com/) - [Examples](https://github.com/mailslurp/examples) repository
 *
 * OpenAPI spec version: 6.5.2
 * Contact: contact@mailslurp.dev
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http.ts';

/**
* IMAP server search options
*/
export class ImapServerSearchOptions {
    'seqNum'?: string;
    'uid'?: string;
    'since'?: Date;
    'before'?: Date;
    'sentSince'?: Date;
    'sentBefore'?: Date;
    'header'?: { [key: string]: Array<string>; };
    'body'?: Array<string>;
    'text'?: Array<string>;
    'withFlags'?: Array<string>;
    'withoutFlags'?: Array<string>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "seqNum",
            "baseName": "seqNum",
            "type": "string",
            "format": ""
        },
        {
            "name": "uid",
            "baseName": "uid",
            "type": "string",
            "format": ""
        },
        {
            "name": "since",
            "baseName": "since",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "before",
            "baseName": "before",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "sentSince",
            "baseName": "sentSince",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "sentBefore",
            "baseName": "sentBefore",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "header",
            "baseName": "header",
            "type": "{ [key: string]: Array<string>; }",
            "format": ""
        },
        {
            "name": "body",
            "baseName": "body",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "text",
            "baseName": "text",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "withFlags",
            "baseName": "withFlags",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "withoutFlags",
            "baseName": "withoutFlags",
            "type": "Array<string>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ImapServerSearchOptions.attributeTypeMap;
    }
    
    public constructor() {
    }
}